Use double-precision height in CmdDraw command-box loop
Holding entry->height as a double across the DrawRect call block matches
the debug target's FPR pressure and codegen, lifting CmdDraw 63.4->63.67
and the unit 70.68->70.74.

+++ src/menu_cmd.cpp
@@ -956,7 +956,7 @@ void CMenuPcs::CmdDraw()
 			const float x = static_cast<float>(entry->x);
 			float y = static_cast<float>(entry->y);
 			const float w = static_cast<float>(entry->width);
-			float h = static_cast<float>(entry->height);
+			double h = static_cast<double>(entry->height);
 			const float u = entry->u;
 			float t = kCmdMenuSmallOffset;
